Specify mirror directories as urls for crusher
Recent changes to spack seem to get confused about identifying
full mirror paths as local files. See, for example, this spack
issue:

https://github.com/spack/spack/issues/34532

An easy solution is to expilcitly specify the path as a file
with the `file://` prefix. These changes are for the configuration
on crusher. Other configurations might also need to be changed.